backend/requirements.txt
# Backend API Requirements
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# HTTP Requests
requests==2.31.0
httpx==0.25.2

# Environment Variables
python-dotenv==1.0.0

# Data Processing
pandas==2.1.3
numpy==1.25.2

# JSON Handling
pydantic==2.5.0

# CORS Support
fastapi-cors==0.0.6

# Additional Utilities
python-multipart==0.0.6
aiofiles==23.2.1

# Angel One SmartAPI
smartapi-python==1.3.0

# Caching
redis==5.0.1
python-memcached==1.62

# Development Tools
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Optional: For enhanced logging
loguru==0.7.2

# Optional: For rate limiting
slowapi==0.1.9

# Optional: For database (if needed later)
# sqlalchemy==2.0.23
# alembic==1.12.1
